v2.0

+++ Product.cs
@@ -1,9 +1,9 @@
 ﻿public class Product
 {
     public string Name { get; private set; }
-    public float Price { get; private set; }
+    public int Price { get; private set; }
     public int Quantity { get; private set; }
-    public Product(string name, float price, int quantity)
+    public Product(string name, int price, int quantity)
     {
         Name = name;
         Price = price;
